Rename repoconfig package to dnfconfig
